KEYCLOAK/app/simple-backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
python-jose[cryptography]
